requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
load-dotenv==0.1.0
lxml==4.9.3
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
slack-sdk==3.26.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
